+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 - TT-2028 - Privacy policy content
 - TT-2026 - Remove terms checkbox
 - TT-2138 - Make nested field labels bold
+- TT-2140 - Add accessibility statement
 
 ### Bug fixes